refactor: Result 이름 변경
