--- src/main/java/racingcar/game/model/Result.java
@@ -1,9 +1,0 @@
-package racingcar.game.model;
-
-import java.util.List;
-import racingcar.car.model.Name;
-
-public interface Result {
-    
-    List<Name> getWinners();
-}
